requirements.txt
Flask==3.1.0
python-docx==1.1.2
PyPDF2==3.0.1
gunicorn==21.2.0
textstat==0.7.3
requests==2.31.0